// src/HyperV.LocalManagement/bin/Release/Publish/types.ts
export type VmEnvironment = 'WMI' | 'HCS';
// ...
export interface Container {
    id: string;
    name: string;
    state: string;
    image: string;
    memoryMB: number;
    cpuCount: number;
    networkMode?: string;
    createdAt?: string;
    startedAt?: string;
}